takako/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.4.2
Django==2.1.4
django-extensions==2.1.4
django-filter==2.0.0
django-filters==0.2.1
django-rest-knox==3.1.3
django-templated-mail==1.1.1
django-webpack-loader==0.6.0
djangorestframework==3.9.0
djoser==1.3.2
idna==2.8
Pillow==5.3.0
pycparser==2.19
pyOpenSSL==18.0.0
pytz==2018.7
six==1.12.0
